requirements.txt
streamlit
python-binance
python-telegram-bot
pandas
python-dotenv
